pyproject.toml
[project]
name = "loop-vault"
version = "0.1.0"
description = ""
authors = [
    {name = "gim-eunhyang",email = "eunhyang90218@gmail.com"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)"
]

[project.optional-dependencies]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
